♻️refactor: 상품 상세 페이지 구조 개선
